+++ experiences/airconditioning/zoneinfo.h
@@ -173,7 +173,7 @@ public:
     };
     ZoneInfos(QObject *parent = nullptr): QAbstractListModel(parent) {}
 
-    int rowCount(const QModelIndex & = QModelIndex()) const override { return m_list.count(); }
+    int rowCount(const QModelIndex & = QModelIndex()) const override { return static_cast<int>(m_list.count()); }
     QVariant data(const QModelIndex &index, int role) const override;
     QHash<int, QByteArray> roleNames() const override;
