+++ libnymea-app/ruletemplates/stateevaluatortemplate.h
@@ -68,7 +68,7 @@ class StateEvaluatorTemplates: public QAbstractListModel
 
 public:
     StateEvaluatorTemplates(QObject *parent = nullptr): QAbstractListModel(parent) {}
-    int rowCount(const QModelIndex &parent = QModelIndex()) const override { Q_UNUSED(parent); return m_list.count(); }
+    int rowCount(const QModelIndex &parent = QModelIndex()) const override { Q_UNUSED(parent); return static_cast<int>(m_list.count()); }
     QVariant data(const QModelIndex &index, int role) const override { Q_UNUSED(index); Q_UNUSED(role); return QVariant(); }
 
     Q_INVOKABLE StateEvaluatorTemplate* get(int index) const {
@@ -80,7 +80,7 @@ public:
 
     void addStateEvaluatorTemplate(StateEvaluatorTemplate *stateEvaluatorTemplate) {
         stateEvaluatorTemplate->setParent(this);
-        beginInsertRows(QModelIndex(), m_list.count(), m_list.count());
+        beginInsertRows(QModelIndex(), static_cast<int>(m_list.count()), static_cast<int>(m_list.count()));
         m_list.append(stateEvaluatorTemplate);
         endInsertRows();
     }
