+++ libnymea-app/types/ruleactionparams.cpp
@@ -34,7 +34,7 @@ RuleActionParams::RuleActionParams(QObject *parent) : QAbstractListModel(parent)
 int RuleActionParams::rowCount(const QModelIndex &parent) const
 {
     Q_UNUSED(parent)
-    return m_list.count();
+    return static_cast<int>(m_list.count());
 }
 
 QVariant RuleActionParams::data(const QModelIndex &index, int role) const
@@ -65,7 +65,7 @@ QHash<int, QByteArray> RuleActionParams::roleNames() const
 void RuleActionParams::addRuleActionParam(RuleActionParam *ruleActionParam)
 {
     ruleActionParam->setParent(this);
-    beginInsertRows(QModelIndex(), m_list.count(), m_list.count());
+    beginInsertRows(QModelIndex(), static_cast<int>(m_list.count()), static_cast<int>(m_list.count()));
     m_list.append(ruleActionParam);
     endInsertRows();
     emit countChanged();
